Moved location code to separate Blueprint
