+++ python_app/app.py
@@ -14,6 +14,7 @@ load_dotenv(dotenv_path)
 from facebookAuth import FbAuth
 from users import Users
 from events import Events, populate_ucla_events_database
+from locations import Locations
 import pymongo
 import json
 
@@ -26,6 +27,7 @@ app = Flask(__name__)
 app.register_blueprint(FbAuth)
 app.register_blueprint(Users)
 app.register_blueprint(Events)
+app.register_blueprint(Locations)
 app.config['SECRET_KEY'] = os.getenv('APP_SECRET_KEY')
 
 # Enable Cross Origin Resource Sharing (CORS)
